+++ Libft/ft_substr.c
@@ -3,11 +3,13 @@
 /*                                                        :::      ::::::::   */
 /*   ft_substr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: martirod <martirod@student.42malaga.com>   +#+  +:+       +#+        */
+/*   By: martirod <martirod@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/08 15:11:43 by martirod          #+#    #+#             */
-/*   Updated: 2024/04/08 15:11:45 by martirod         ###   ########.fr       */
+/*   Updated: 2024/04/15 18:52:24 by martirod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-char *ft_substr(char const *s, unsigned int start, size_t len);
+#include "libft.h"
+
+char	*ft_substr(char const *s, size_t start, size_t len);
